[Refactor] favorite 관련 매장 조회 로직 market 패키지로 옮김 - #54
