+++ src/main/java/com/appcenter/marketplace/domain/favorite/repository/FavoriteRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface FavoriteRepository extends JpaRepository<Favorite,Long>, FavoriteRepositoryCustom{
+public interface FavoriteRepository extends JpaRepository<Favorite,Long>{
 
     Optional<Favorite> findByMember_IdAndMarket_Id(Long memberId, Long marketId);
