--- src/main/java/com/appcenter/marketplace/domain/favorite/repository/FavoriteRepositoryCustom.java
@@ -1,12 +1,0 @@
-package com.appcenter.marketplace.domain.favorite.repository;
-
-import com.appcenter.marketplace.domain.market.dto.res.MarketResDto;
-
-import java.util.List;
-
-public interface FavoriteRepositoryCustom {
-
-    public List<MarketResDto> findFavoriteMarketResDtoByMemberId(Long memberId, Long marketId, Integer size);
-
-    public List<MarketResDto> findTopFavoriteMarketResDto(Long marketId, Integer size);
-}
