+++ src/main/java/com/appcenter/marketplace/domain/favorite/service/impl/FavoriteServiceImpl.java
@@ -4,8 +4,6 @@ import com.appcenter.marketplace.domain.favorite.Favorite;
 import com.appcenter.marketplace.domain.favorite.repository.FavoriteRepository;
 import com.appcenter.marketplace.domain.favorite.service.FavoriteService;
 import com.appcenter.marketplace.domain.market.Market;
-import com.appcenter.marketplace.domain.market.dto.res.MarketPageResDto;
-import com.appcenter.marketplace.domain.market.dto.res.MarketResDto;
 import com.appcenter.marketplace.domain.market.repository.MarketRepository;
 import com.appcenter.marketplace.domain.member.Member;
 import com.appcenter.marketplace.domain.member.MemberRepository;
@@ -15,7 +13,6 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
 import java.util.Optional;
 
 @Transactional(readOnly = true)
@@ -46,31 +43,6 @@ public class FavoriteServiceImpl implements FavoriteService {
         }
     }
 
-    @Override
-    public MarketPageResDto getMemberFavoriteMarketPage(Long memberId, Long marketId, Integer size) {
-        List<MarketResDto> marketResDtoList=favoriteRepository.findFavoriteMarketResDtoByMemberId(memberId,marketId,size);
-
-        return checkHasNextPageAndReturnPageDto(marketResDtoList,size);
-    }
-
-    @Override
-    public MarketPageResDto getTopFavoriteMarketPage(Long marketId, Integer size) {
-        List<MarketResDto> marketResDtoList=favoriteRepository.findTopFavoriteMarketResDto(marketId,size);
-
-        return checkHasNextPageAndReturnPageDto(marketResDtoList,size);
-    }
-
-    private MarketPageResDto checkHasNextPageAndReturnPageDto(List<MarketResDto> marketResDtoList, Integer size){
-        boolean hasNext=false;
-
-        // 가져온 갯수가 페이지 사이즈보다 많으면 다음 페이지가 있는 것이고, 사이즈에 맞게 조정한다.
-        if(marketResDtoList.size()>size){
-            hasNext=true;
-            marketResDtoList.remove(size.intValue());
-        }
-
-        return new MarketPageResDto(marketResDtoList,hasNext);
-    }
 
 }
