+++ src/main/java/com/appcenter/marketplace/domain/market/repository/MarketRepositoryCustom.java
@@ -11,4 +11,8 @@ public interface MarketRepositoryCustom {
     public List<MarketResDto> findMarketResDtoList(Long marketId, Integer size);
 
     public List<MarketResDto> findMarketResDtoListByCategory(Long marketId, Integer size, String major);
+
+    public List<MarketResDto> findFavoriteMarketResDtoByMemberId(Long memberId, Long marketId, Integer size);
+
+    public List<MarketResDto> findTopFavoriteMarketResDto(Long marketId, Integer size);
 }
