+++ src/main/java/com/appcenter/marketplace/domain/market/repository/MarketRepositoryCustomImpl.java
@@ -12,6 +12,7 @@ import lombok.RequiredArgsConstructor;
 import java.util.List;
 
 import static com.appcenter.marketplace.domain.category.QCategory.category;
+import static com.appcenter.marketplace.domain.favorite.QFavorite.favorite;
 import static com.appcenter.marketplace.domain.image.QImage.image;
 import static com.appcenter.marketplace.domain.market.QMarket.market;
 import static com.querydsl.core.group.GroupBy.groupBy;
@@ -70,6 +71,31 @@ public class MarketRepositoryCustomImpl implements MarketRepositoryCustom{
                 .orderBy(market.id.desc())
                 .limit(size + 1)
                 .fetch();
+    }
+
+    @Override
+    public List<MarketResDto> findFavoriteMarketResDtoByMemberId(Long memberId, Long marketId, Integer size) {
+        return jpaQueryFactory
+                .select(new QMarketResDto(market.id, market.name, market.description, market.thumbnail))
+                .from(market)
+                .innerJoin(favorite).on(market.eq(favorite.market))
+                .where(ltMarketId(marketId).and(favorite.member.id.eq(memberId)).and(favorite.isDeleted.eq(false)))
+                .orderBy(favorite.modifiedAt.desc())
+                .limit(size + 1)
+                .fetch();
+    }
+
+    @Override
+    public List<MarketResDto> findTopFavoriteMarketResDto(Long marketId, Integer size) {
+        return jpaQueryFactory
+                .select(new QMarketResDto(market.id, market.name, market.description, market.thumbnail))
+                .from(market)
+                .leftJoin(favorite).on(market.eq(favorite.market)) // market과 favorite를 조인
+                .where(ltMarketId(marketId).and(favorite.isDeleted.eq(false))) // 삭제되지 않은 찜만 필터링
+                .groupBy(market.id) // 매장별로 그룹화
+                .orderBy(favorite.member.id.count().desc()) // 찜 수가 많은 순으로 정렬
+                .limit(size+1) // 반환할 리스트 크기 제한
+                .fetch(); // 결과 반환
     }
 
     // BooleanBuilder은 QueryDSL에서 조건을 표현하는 객체이다. 빈 객체를 반환하면 아무런 조건없이 실행된다.
