+++ src/main/java/com/appcenter/marketplace/domain/market/service/MarketService.java
@@ -8,4 +8,8 @@ public interface MarketService {
     MarketDetailsResDto getMarketDetails(Long marketId);
 
     MarketPageResDto getMarketPage(Long marketId, Integer size, String major);
+
+    MarketPageResDto getMemberFavoriteMarketPage(Long memberId, Long marketId, Integer size);
+
+    MarketPageResDto getTopFavoriteMarketPage(Long marketId, Integer size);
 }
